Version 1.06

+++ src/main/java/pl/edu/wszib/controller/BuyController.java
@@ -46,6 +46,9 @@ public class BuyController {
 
     @RequestMapping("/buy")
     public String vehicle(Model model, Vehicle vehicle, BindingResult bindingResult){
+        if(this.sessionObject.getUser() == null){
+            return "redirect:login";
+        }
         if(bindingResult.hasErrors()) {
             return "buy"; // powrót do formularza
         }
@@ -57,6 +60,9 @@ public class BuyController {
 
     @GetMapping("/buy")
     public String VehicleForm(Vehicle vehicle) {
+        if(this.sessionObject.getUser() == null){
+            return "redirect:login";
+        }
         return "buy";
     }
 
@@ -69,6 +75,9 @@ public class BuyController {
 
     @RequestMapping(value = "/buy2",method = RequestMethod.POST)
     public String addVehicleForm(@ModelAttribute("getBuyKey") Buy buy){
+        if(this.sessionObject.getUser() == null){
+            return "redirect:login";
+        }
         buyDAO.addBuy(buy);
         return  "redirect:resultBuy";
     }
@@ -80,13 +89,18 @@ public class BuyController {
 
     @RequestMapping(value = "/resultBuy",method = RequestMethod.POST)
     public String removeVehicleForm(@ModelAttribute("BuyKey") Vehicle vehicle){
+        if(this.sessionObject.getUser() == null){
+            return "redirect:login";
+        }
         vehicleDAO.removeVehicleId(vehicle.getIdVehicle());
         return "redirect:thanks";
     }
 
     @RequestMapping(value = "/thanks",method = RequestMethod.GET)
     public String indexGetVehicle(){
-
+        if(this.sessionObject.getUser() == null){
+            return "redirect:login";
+        }
         return"thanks";
     }
 }
